CRUD operation for structure done

+++ src/main/java/com/example/dealsplus/service/StructureService.java
@@ -3,10 +3,8 @@ package com.example.dealsplus.service;
 import com.example.dealsplus.dto.StructureDto;
 import com.example.dealsplus.exception.CustomException;
 import com.example.dealsplus.model.Structure;
-import com.example.dealsplus.model.User;
 import com.example.dealsplus.repo.StructureRepo;
-import com.example.dealsplus.repo.UserRepo;
-import org.modelmapper.ModelMapper;
+import com.example.dealsplus.utils.ConstantUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,19 +15,13 @@ import java.util.stream.Collectors;
 public class StructureService {
 
     private final StructureRepo structureRepo;
-    private final UserRepo userRepo;
-    private final ModelMapper modelMapper;
+    private final UserService userService;
 
-    public StructureService(StructureRepo structureRepo, UserRepo userRepo, ModelMapper modelMapper) {
+    public StructureService(StructureRepo structureRepo, UserService userService) {
         this.structureRepo = structureRepo;
-        this.userRepo = userRepo;
-        this.modelMapper = modelMapper;
+        this.userService = userService;
     }
 
-    public void createService(StructureDto structure) {
-        Structure map = new Structure(structure.getStructureName());
-        saveStructure(map);
-    }
 
     public Structure getStructure(String structureName) {
         Optional<Structure> optionalStructure = structureRepo.findByStructureName(structureName);
@@ -39,28 +31,38 @@ public class StructureService {
         return optionalStructure.get();
     }
 
-    public void saveStructure(Structure structure) {
-        structureRepo.save(structure);
+    public StructureDto getStructureDto(Structure structure){
+        return new StructureDto(structure.getStructureName(), structure.getStructureInfo());
     }
 
-    public void deleteStructure(Structure structure) {
+    public String createService(StructureDto structureDto) {
+        userService.checkForAuthorization(ConstantUtils.createPermissionForEntity.apply(ConstantUtils.STRUCTURE_ENTITY, ConstantUtils.CREATE_PERM_STRING));
+        Structure structure = new Structure(structureDto.getStructureName());
+        structureRepo.save(structure);
+        return structureDto.getStructureName();
+    }
+
+    public void deleteStructure(String structureName) {
+        userService.checkForAuthorization(ConstantUtils.createPermissionForEntity.apply(ConstantUtils.STRUCTURE_ENTITY , ConstantUtils.DELETE_PERM_STRING));
+        Structure structure = getStructure(structureName);
         structureRepo.delete(structure);
     }
 
     public StructureDto readStructure(String structureName) {
+        userService.checkForAuthorization(ConstantUtils.createPermissionForEntity.apply(ConstantUtils.STRUCTURE_ENTITY , ConstantUtils.READ_PERM_STRING));
         Structure structure = getStructure(structureName);
-        return new StructureDto(structure.getStructureName());
-    }
-
-    private Optional<User> getUser(String username) {
-        Optional<User> optionalUser = userRepo.findByUsername(username);
-        if (optionalUser.isEmpty()) {
-            throw new CustomException("User not present");
-        }
-        return optionalUser;
+        return new StructureDto(structure.getStructureName(), structure.getStructureInfo());
     }
 
     public List<String> getAll() {
+        userService.checkForAuthorization(ConstantUtils.createPermissionForEntity.apply(ConstantUtils.STRUCTURE_ENTITY , ConstantUtils.READ_PERM_STRING));
         return structureRepo.findAll().stream().map(Structure::getStructureName).collect(Collectors.toList());
+    }
+
+    public void updateStructure(StructureDto structureDto) {
+        userService.checkForAuthorization(ConstantUtils.createPermissionForEntity.apply(ConstantUtils.STRUCTURE_ENTITY , ConstantUtils.UPDATE_PERM_STRING));
+        Structure structure = getStructure(structureDto.getStructureName());
+        structure.setStructureInfo(structureDto.getStructureInfo());
+        structureRepo.save(structure);
     }
 }
